# api/openapi/agent-control.v1.yaml
openapi: 3.1.0
info:
  title: MiviaLabs Agent Control API
  version: 0.1.0
  description: Localhost-only bootstrap REST contract for task, research-run, project context, reliability, and redacted agent-run metadata.
...
servers:
  - url: http://127.0.0.1:8080/api/v1
paths:
...
  /projects/{id}/work-plans:
    post:
      operationId: createProjectWorkPlan
      summary: Create project Work Plan
      description: Creates one project-scoped metadata-only Work Plan. Raw prompts, completions, source dumps, raw stderr, provider payloads, secrets, roots, external URLs, skipped sensitive content, and PII are prohibited.
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectWorkPlanRequest'
      responses:
        '201':
          description: Created Work Plan
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectWorkPlan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
    get:
      operationId: listProjectWorkPlans
      summary: List project Work Plans
      description: Lists project-scoped Work Plan metadata with bounded pagination and safe metadata filters.
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/WorkPlanStatusQuery'
        - $ref: '#/components/parameters/OwnerAgentQuery'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageToken'
      responses:
        '200':
          description: Work Plans
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectWorkPlanList'
        '400':
          $ref: '#/components/responses/BadRequest'
  /projects/{id}/work-plans/{plan_id}:
    get:
      operationId: getProjectWorkPlan
      summary: Get project Work Plan
      description: Returns one project-scoped Work Plan metadata record.
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/PlanID'
      responses:
        '200':
          description: Work Plan
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectWorkPlan'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{id}/work-plans/{plan_id}/status:
    post:
      operationId: updateProjectWorkPlanStatus
      summary: Update project Work Plan status
      description: Updates Work Plan status and resume metadata without storing raw prompts, completions, source dumps, raw stderr, provider payloads, secrets, roots, external URLs, or PII.
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/PlanID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectWorkPlanStatusRequest'
      responses:
        '200':
          description: Updated Work Plan
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectWorkPlan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /projects/{id}/work-plans/{plan_id}/resume:
    get:
      operationId: resumeProjectWorkPlan
      summary: Resume project Work Plan
      description: Returns metadata needed to resume a Work Plan, including safe task refs, dependency summaries, context refs, and verification requirements.
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/PlanID'
      responses:
        '200':
          description: Work Plan resume metadata
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectWorkPlan'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{id}/work-plans/{plan_id}/tasks:
    post:
      operationId: createProjectWorkTask
      summary: Create project Work Task
      description: Creates one isolated-worker-ready project Work Task under a Work Plan. Tasks are metadata-only and must include bounded scope, dependencies, evidence needs, context refs, verification requirement, and resume instructions. Each task must be executable by a low-intelligence worker from task metadata and attached refs alone, without prior chat memory or hidden orchestrator context; verification must be written for independent orchestrator execution.
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/PlanID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectWorkTaskRequest'
      responses:
        '201':
          description: Created Work Task
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectWorkTask'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /projects/{id}/work-tasks:
    get:
      operationId: listProjectWorkTasks
      summary: List project Work Tasks
      description: Lists project-scoped Work Task metadata with bounded pagination and safe metadata filters.
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/PlanIDQuery'
        - $ref: '#/components/parameters/WorkTaskStatusQuery'
        - $ref: '#/components/parameters/OwnerAgentQuery'
        - $ref: '#/components/parameters/ClaimedByRunIDQuery'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageToken'
      responses:
        '200':
          description: Work Tasks
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectWorkTaskList'
        '400':
          $ref: '#/components/responses/BadRequest'
  /projects/{id}/work-tasks/open:
    get:
      operationId: listOpenProjectWorkTasks
      summary: List open project Work Tasks
      description: Lists ready, claimed, in-progress, blocked, needs-review, verifying, and planned Work Tasks according to the service's open-task rules.
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/PlanIDQuery'
        - $ref: '#/components/parameters/WorkTaskStatusQuery'
        - $ref: '#/components/parameters/OwnerAgentQuery'
        - $ref: '#/components/parameters/ClaimedByRunIDQuery'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageToken'
      responses:
        '200':
          description: Open Work Tasks
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectWorkTaskList'
        '400':
          $ref: '#/components/responses/BadRequest'
  /projects/{id}/work-tasks/mine:
    get:
      operationId: listMyProjectWorkTasks
      summary: List Work Tasks claimed by the current run
      description: Lists project Work Tasks filtered by claimed_by_run_id or run_id.
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/PlanIDQuery'
        - $ref: '#/components/parameters/WorkTaskStatusQuery'
        - $ref: '#/components/parameters/OwnerAgentQuery'
        - $ref: '#/components/parameters/ClaimedByRunIDQuery'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageToken'
      responses:
        '200':
          description: Claimed Work Tasks
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectWorkTaskList'
        '400':
          $ref: '#/components/responses/BadRequest'
  /projects/{id}/work-tasks/blocked:
    get:
      operationId: listBlockedProjectWorkTasks
      summary: List blocked project Work Tasks
      description: Lists blocked project-scoped Work Task metadata.
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/PlanIDQuery'
        - $ref: '#/components/parameters/OwnerAgentQuery'
        - $ref: '#/components/parameters/ClaimedByRunIDQuery'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageToken'
      responses:
        '200':
          description: Blocked Work Tasks
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectWorkTaskList'
        '400':
          $ref: '#/components/responses/BadRequest'
  /projects/{id}/work-tasks/next:
    get:
      operationId: getNextProjectWorkTask
      summary: Get next project Work Task
      description: Selects the next claimable Work Task using safe metadata, dependency readiness, context refs, and verification requirements.
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/PlanIDQuery'
        - $ref: '#/components/parameters/OwnerAgentQuery'
        - name: run_id
          in: query
          schema:
            type: string
            maxLength: 200
        - name: trace_id
          in: query
          schema:
            type: string
            maxLength: 200
        - name: include_claimed_by_me
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Next Work Task selection
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectWorkTaskNext'
        '400':
          $ref: '#/components/responses/BadRequest'
  /projects/{id}/work-tasks/{task_id}:
    get:
      operationId: getProjectWorkTask
      summary: Get project Work Task
      description: Returns one project-scoped Work Task metadata record.
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/TaskID'
      responses:
        '200':
          description: Work Task
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectWorkTask'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{id}/work-tasks/{task_id}/claim:
    post:
      operationId: claimProjectWorkTask
      summary: Claim project Work Task
      description: Claims one Work Task for a run using metadata-only refs.
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/TaskID'
      requestBody:
        $ref: '#/components/requestBodies/ProjectWorkTaskAction'
      responses:
        '200':
          description: Claimed Work Task
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectWorkTask'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /projects/{id}/work-tasks/{task_id}/release:
    post:
      operationId: releaseProjectWorkTask
      summary: Release project Work Task
      description: Releases a claimed Work Task without deleting task metadata.
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/TaskID'
      requestBody:
        $ref: '#/components/requestBodies/ProjectWorkTaskAction'
      responses:
        '200':
          description: Released Work Task
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectWorkTask'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /projects/{id}/work-tasks/{task_id}/start:
    post:
      operationId: startProjectWorkTask
      summary: Start project Work Task
      description: Marks a claimed Work Task in progress and links safe run metadata.
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/TaskID'
      requestBody:
        $ref: '#/components/requestBodies/ProjectWorkTaskAction'
      responses:
        '200':
          description: Started Work Task
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectWorkTask'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
...
  /projects/{id}/work-tasks/{task_id}/complete:
    post:
      operationId: completeProjectWorkTask
      summary: Complete project Work Task
      description: Completes a Work Task with bounded outcome text and verifier result refs.
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/TaskID'
      requestBody:
        $ref: '#/components/requestBodies/ProjectWorkTaskAction'
      responses:
        '200':
          description: Completed Work Task
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectWorkTask'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
...
components:
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
  headers:
    NoStore:
      description: Responses are metadata-only and are not intended for intermediary caching.
      schema:
        type: string
        enum: [no-store]
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    CreateProjectWorkPlanRequest:
      type: object
      additionalProperties: false
      required: [plan_ref, title, goal_summary]
      properties:
        plan_ref: {type: string, minLength: 1, maxLength: 200}
        user_request_ref: {type: string, maxLength: 200}
        title: {type: string, minLength: 1, maxLength: 200}
        goal_summary: {type: string, minLength: 1, maxLength: 500}
        owner_agent: {type: string, maxLength: 120}
        created_by_run_id: {type: string, maxLength: 200}
        trace_id: {type: string, maxLength: 200}
        resume_summary: {type: string, maxLength: 500}
        isolation_mode:
          type: string
          enum: [shared, dedicated_worktree, unavailable]
        parallel_group_ref: {type: string, maxLength: 200}
        workspace_ref: {type: string, maxLength: 200}
        git_base_ref: {type: string, maxLength: 200}
        git_branch_ref: {type: string, maxLength: 200}
        git_worktree_ref: {type: string, maxLength: 200}
      description: Metadata-only request. Do not include raw prompts, completions, source dumps, raw stderr, provider payloads, secrets, roots, external URLs, skipped sensitive content, or PII.
...